+++ packages/orq-rc/src/models/operations/findonepromptsnippet.ts
@@ -4322,7 +4322,7 @@ export const FindOnePromptSnippetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-02-04T10:25:39.047Z",
+    "2025-02-04T10:49:03.956Z",
   ).transform(v => new Date(v)),
   type: FindOnePromptSnippetType$inboundSchema,
   versions: z.array(z.lazy(() => FindOnePromptSnippetVersions$inboundSchema)),
@@ -4372,7 +4372,7 @@ export const FindOnePromptSnippetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string(),
   updatedById: z.string(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-02-04T10:25:39.047Z"))
+  updated: z.date().default(() => new Date("2025-02-04T10:49:03.956Z"))
     .transform(v => v.toISOString()),
   type: FindOnePromptSnippetType$outboundSchema,
   versions: z.array(z.lazy(() => FindOnePromptSnippetVersions$outboundSchema)),